requirements.txt
GitPython==3.1.45
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pip==25.2
plotly==6.3.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
streamlit==1.49.1